// packages/chapterjs/src/discord/types/message.ts
export const MessageFlags = {
  /** this message has been published to subscribed channels (via Channel Following) */
  Crossposted: 1 << 0,
  /** this message originated from a message in another channel (via Channel Following) */
  IsCrosspost: 1 << 1,
  /** do not include any embeds when serializing this message */
  SuppressEmbeds: 1 << 2,
  /** the source message for this crosspost has been deleted (via Channel Following) */
  SourceMessageDeleted: 1 << 3,
  /** this message came from the urgent message system */
  Urgent: 1 << 4,
  /** this message has an associated thread, with the same id as the message */
  HasThread: 1 << 5,
  /** this message is only visible to the user who invoked the Interaction */
  Ephemeral: 1 << 6,
  /** this message is an Interaction Response and the bot is "thinking" */
  Loading: 1 << 7,
  /** this message failed to mention some roles and add their members to the thread */
  FailedToMentionSomeRolesInThread: 1 << 8,
  /** this message will not trigger push and desktop notifications */
  SuppressNotifications: 1 << 12,
  /** this message is a voice message */
  IsVoiceMessage: 1 << 13,
  /** this message has a snapshot (via Message Forwarding) */
  HasSnapshot: 1 << 14,
  /** allows you to create fully component-driven messages */
  IsComponentsV2: 1 << 15,
} as const;
export type MessageFlags = (typeof MessageFlags)[keyof typeof MessageFlags];
// ...
export interface RawMessageReference {
  /** type of reference. */
  type?: MessageReferenceType;
  /** id of the originating message */
  message_id?: Snowflake;
  /** id of the originating message's channel */
  channel_id?: Snowflake;
  /** id of the originating message's guild */
  guild_id?: Snowflake;
  /** when sending, whether to error if the referenced message doesn't exist instead of sending as a normal (non-reply) message, default true */
  fail_if_not_exists?: boolean;
}

/**
 * Message Reference Types
 * @see https://docs.discord.com/developers/resources/message#message-reference-types
 */
export const MessageReferenceType = {
  /** A standard reference used by replies. */
  Default: 0,
  /** Reference used to point to a message at a point in time. */
  Forward: 1,
} as const;
export type MessageReferenceType =
  (typeof MessageReferenceType)[keyof typeof MessageReferenceType];
// ...
export interface RawEmbed {
  /** title of embed */
  title?: string;
  /** type of embed (always "rich" for webhook embeds) */
  type?: EmbedType;
  /** description of embed */
  description?: string;
  /** url of embed */
  url?: string;
  /** timestamp of embed content */
  timestamp?: ISO8601Timestamp;
  /** color code of the embed */
  color?: number;
  /** footer information */
  footer?: RawEmbedFooter;
  /** image information */
  image?: RawEmbedImage;
  /** thumbnail information */
  thumbnail?: RawEmbedImage;
  /** video information */
  video?: RawEmbedVideo;
  /** provider information */
  provider?: RawEmbedProvider;
  /** author information */
  author?: RawEmbedAuthor;
  /** fields information, max of 25 */
  fields?: RawEmbedField[];
  /** embed flags combined as a bitfield */
  flags?: number;
}
// ...
export interface RawEmbedVideo {
  /** source url of video */
  url?: string;
  /** a proxied url of the video */
  proxy_url?: string;
  /** height of video */
  height?: number;
  /** width of video */
  width?: number;
  /** the video's media type */
  content_type?: string;
  /** thumbhash placeholder of the video */
  placeholder?: string;
  /** version of the placeholder */
  placeholder_version?: number;
  /** description (alt text) for the video */
  description?: string;
  /** embed media flags combined as a bitfield */
  flags?: number;
}

/**
 * Embed Image Structure
 * @see https://docs.discord.com/developers/resources/message#embed-object-embed-image-structure
 */
export interface RawEmbedImage {
  /** source url of image (only supports http(s) and attachments) */
  url: string;
  /** a proxied url of the image */
  proxy_url?: string;
  /** height of image */
  height?: number;
  /** width of image */
  width?: number;
  /** the image's media type */
  content_type?: string;
  /** thumbhash placeholder of the image */
  placeholder?: string;
  /** version of the placeholder */
  placeholder_version?: number;
  /** description (alt text) for the image */
  description?: string;
  /** embed media flags combined as a bitfield */
  flags?: number;
}
// ...
export interface RawEmbedProvider {
  /** name of provider */
  name?: string;
  /** url of provider */
  url?: string;
}

/**
 * Embed Author Structure
 * @see https://docs.discord.com/developers/resources/message#embed-object-embed-author-structure
 */
export interface RawEmbedAuthor {
  /** name of author */
  name: string;
  /** url of author (only supports http(s)) */
  url?: string;
  /** url of author icon (only supports http(s) and attachments) */
  icon_url?: string;
  /** a proxied url of author icon */
  proxy_icon_url?: string;
}

/**
 * Embed Footer Structure
 * @see https://docs.discord.com/developers/resources/message#embed-object-embed-footer-structure
 */
export interface RawEmbedFooter {
  /** footer text */
  text: string;
  /** url of footer icon (only supports http(s) and attachments) */
  icon_url?: string;
  /** a proxied url of footer icon */
  proxy_icon_url?: string;
}

/**
 * Embed Field Structure
 * @see https://docs.discord.com/developers/resources/message#embed-object-embed-field-structure
 */
export interface RawEmbedField {
  /** name of the field */
  name: string;
  /** value of the field */
  value: string;
  /** whether or not this field should display inline */
  inline?: boolean;
}
// ...
export interface RawAttachmentRequest {
  /** attachment id, for new attachments this must match the `n` in `files[n]` */
  id: Snowflake | number;
  /** name of file attached */
  filename?: string;
  /** the title of the file */
  title?: string;
  /** description (alt text) for the file (max 1024 characters) */
  description?: string;
  /** the duration of the audio or video file (required for voice messages) */
  duration_secs?: number;
  /** base64 encoded bytearray representing a sampled waveform (required for voice messages) */
  waveform?: string;
  /** whether the attachment should be marked as a spoiler and blurred until clicked, this sets the `IS_SPOILER` attachment flag */
  is_spoiler?: boolean;
}
// ...
export const AllowedMentionType = {
  /** Controls role mentions */
  Roles: 'roles',
  /** Controls user mentions */
  Users: 'users',
  /** Controls @everyone and @here mentions */
  Everyone: 'everyone',
} as const;
export type AllowedMentionType =
  (typeof AllowedMentionType)[keyof typeof AllowedMentionType];

/**
 * Default Settings for Allowed Mentions
 * @see https://docs.discord.com/developers/resources/message#allowed-mentions-object-default-settings-for-allowed-mentions
 */
export interface RawAllowedMentions {
  /** An array of allowed mention types to parse from the content */
  parse?: AllowedMentionType[];
  /** Array of role ids to mention, max 100 */
  roles?: Snowflake[];
  /** Array of user ids to mention, max 100 */
  users?: Snowflake[];
  /** For replies, whether to mention the author of the message being replied to, defaults to false */
  replied_user?: boolean;
}
// ...
export interface RawSharedClientTheme {
  /** the hexadecimal-encoded colors of the theme (max of 5) */
  colors: string[];
  /** the direction of the theme's colors (max of 360) */
  gradient_angle: number;
  /** the intensity of the theme's colors (max of 100) */
  base_mix: number;
  /** the mode of the theme */
  base_theme?: SharedClientThemeBaseType | null;
}

/**
 * Base Theme Types
 * @see https://docs.discord.com/developers/resources/message#base-theme-types
 */
export const SharedClientThemeBaseType = {
  Unset: 0,
  Dark: 1,
  Light: 2,
  Darker: 3,
  Midnight: 4,
} as const;
export type SharedClientThemeBaseType =
  (typeof SharedClientThemeBaseType)[keyof typeof SharedClientThemeBaseType];
// ...
export interface CreateMessageJSONParams {
  /** Message contents (up to 2000 characters) */
  content?: string;
  /** Can be used to verify a message was sent (up to 25 characters). Value will appear in the Message Create event. */
  nonce?: number | string;
  /** `true` if this is a TTS message */
  tts?: boolean;
  /** Up to 10 `rich` embeds (up to 6000 characters) */
  embeds?: RawEmbed[];
  /** Allowed mentions for the message */
  allowed_mentions?: RawAllowedMentions;
  /** Include to make your message a reply or a forward */
  message_reference?: RawMessageReference;
  /** Components to include with the message */
  components?: RawComponent[];
  /** IDs of up to 3 stickers in the server to send in the message */
  sticker_ids?: Snowflake[];
  /** JSON-encoded body of non-file params, only for `multipart/form-data` requests. See Uploading Files */
  payload_json?: string;
  /** Metadata for the attachments. See Uploading Files */
  attachments?: Partial<RawAttachmentRequest>[];
  /** Message flags combined as a bitfield (only `SUPPRESS_EMBEDS`, `SUPPRESS_NOTIFICATIONS`, `IS_VOICE_MESSAGE`, and `IS_COMPONENTS_V2` can be set) */
  flags?: number;
  /** If true and nonce is present, it will be checked for uniqueness in the past few minutes. If another message was created by the same author with the same nonce, that message will be returned and no new message will be created. */
  enforce_nonce?: boolean;
  /** A poll! */
  poll?: RawPollCreateRequest;
  /** The custom client-side theme to share via the message */
  shared_client_theme?: RawSharedClientTheme;
}
// ...
export type EmbedType =
  'rich' | 'image' | 'video' | 'gifv' | 'article' | 'link' | 'poll_result';